+++ src/handler/HandlerTracksGO.h
@@ -6,7 +6,7 @@ class HandlerTracksGO : public HandlerGO
 {
 public:
     HandlerTracksGO(Particle *particle, Light *incidentLight, int nTheta,
-                    float wavelength);
+                    double wavelength);
 
     void HandleBeams(std::vector<Beam> &beams, double sinZenith) override;
     void WriteMatricesToFile(std::string &destName, double nrg) override;
